pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]

[project]
name = "maja-mag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.4",
    "build>=1.2.2.post1",
    "dotenv>=0.9.9",
    "kaleido>=1.0.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.1",
    "mediapipe>=0.10.14",
    "mlflow>=3.1.0",
    "moviepy==1.0.3",
    "numpy==1.26.3",
    "onnx>=1.18.0",
    "onnxruntime>=1.20.1",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python-headless>=4.11.0.86",
    "openmim>=0.3.9",
    "optuna>=4.4.0",
    "pandas>=2.2.3",
    "plotly>=6.2.0",
    "pydantic-settings>=2.9.1",
    "pyinstaller>=6.13.0",
    "pyqt6>=6.9.0",
    "pytorch-lightning>=2.5.1.post0",
    "pyyaml>=6.0.2",
    "ruff>=0.9.8",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "setuptools>=80.3.1",
    "shap>=0.48.0",
    "tk>=0.1.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "twine>=6.1.0",
    "ultralytics>=8.3.142",
    "xgboost>=3.0.4",
]


[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "B", "ANN"]

[tool.ruff.lint.isort]
lines-between-types = 1

[tool.setuptools]
package-dir = {"" = "src"}
